requirements.txt
flask==2.3.3
flask-socketio==5.3.6
paho-mqtt==1.6.1
python-socketio==5.9.0
python-engineio==4.8.0
eventlet==0.33.3
